+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Update Airplay icon to modern variant
   * Add gradient to background to break up the solid color
   * Reformat Admin Settings portion of Admin Panel (previously known as the Updater)
+  * Add ability to create streams, presets from the home screen
 
 ## 0.4.5
 * Web App